Store the last window location and restore it on the next start
This fixes an annoying problem with a multi-monitor setup where the
window was shown at the center of the combined screen, e.g. in case
of two screens with identical resolution the window was shown with left
half on right side of left monitor and with right half on the left side
of the right monitor. This is a known bug in the PySimpleGUI. This still
happens on the first start, but on subsequent starts the last remembered
window location is restored. Unfortunately, we cannot remember the
location just before the window is closed because window.read() returns
after the window is already closed.

Note: After reading the last location from the settings we have to
convert it explicitly to a tuple because Window() expects a tuple, but
UserSettings returns a list (because JSON only knows lists).